LPS-54821 Remove UpgradeAssetPublisherManualEntries class and move methods into appropirate upgrade class.
