+++ portal-impl/src/com/liferay/portal/upgrade/v6_0_0/UpgradeAssetPublisher.java
@@ -15,14 +15,20 @@
 package com.liferay.portal.upgrade.v6_0_0;
 
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
+import com.liferay.portal.kernel.util.ArrayUtil;
 import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
-import com.liferay.portal.upgrade.util.UpgradeAssetPublisherManualEntries;
 import com.liferay.portlet.PortletPreferencesFactoryUtil;
+import com.liferay.portlet.journal.model.JournalArticle;
+import com.liferay.portlet.journal.model.JournalArticleResource;
+import com.liferay.portlet.journal.service.persistence.JournalArticleResourceUtil;
+import com.liferay.portlet.journal.service.persistence.JournalArticleUtil;
+
+import java.util.List;
 
 import javax.portlet.PortletPreferences;
 
@@ -31,6 +37,36 @@ import javax.portlet.PortletPreferences;
  * @author Douglas Wong
  */
 public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
+
+	public static void upgradeToAssetEntryIdElement(Element rootElement) {
+		Element assetIdElement = rootElement.element("asset-id");
+
+		if (assetIdElement != null) {
+			String assetEntryId = assetIdElement.getText();
+
+			Element assetEntryIdElement = rootElement.addElement(
+				"assetEntryId");
+
+			assetEntryIdElement.addText(assetEntryId);
+
+			rootElement.remove(assetIdElement);
+		}
+	}
+
+	public static void upgradeToAssetEntryTypeElement(Element rootElement) {
+		Element assetTypeElement = rootElement.element("asset-type");
+
+		if (assetTypeElement != null) {
+			String assetEntryType = assetTypeElement.getText();
+
+			Element assetEntryTypeElement = rootElement.addElement(
+				"assetEntryType");
+
+			assetEntryTypeElement.addText(assetEntryType);
+
+			rootElement.remove(assetTypeElement);
+		}
+	}
 
 	protected String[] getAssetEntryXmls(String[] manualEntries)
 		throws Exception {
@@ -44,11 +80,9 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 
 			Element rootElement = document.getRootElement();
 
-			UpgradeAssetPublisherManualEntries.upgradeToAssetEntryIdElement(
-				rootElement);
+			upgradeToAssetEntryIdElement(rootElement);
 
-			UpgradeAssetPublisherManualEntries.upgradeToAssetEntryTypeElement(
-				rootElement);
+			upgradeToAssetEntryTypeElement(rootElement);
 
 			assetEntryXmls[i] = document.formattedString(StringPool.BLANK);
 		}
@@ -59,6 +93,19 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 	@Override
 	protected String[] getPortletIds() {
 		return new String[] {"101_INSTANCE_%"};
+	}
+
+	@Override
+	protected String getUpdatePortletPreferencesWhereClause() {
+		StringBundler sb = new StringBundler(5);
+
+		sb.append("(portletId like '101_INSTANCE_%') and ((preferences like ");
+		sb.append("'%<preference><name>selection-style</name><value>manual");
+		sb.append("</value></preference>%') OR (preferences like ");
+		sb.append("'%<preference><name>selectionStyle</name><value>manual");
+		sb.append("</value></preference>%'))");
+
+		return sb.toString();
 	}
 
 	@Override
@@ -148,7 +195,71 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 			portletPreferences.setValues("asset-entry-xml", assetEntryXmls);
 		}
 
+		String[] assetEntryXmls = portletPreferences.getValues(
+			"asset-entry-xml", new String[0]);
+
+		if (ArrayUtil.isEmpty(assetEntryXmls)) {
+			assetEntryXmls = portletPreferences.getValues(
+				"assetEntryXml", new String[0]);
+		}
+
+		String[] manualEntries = portletPreferences.getValues(
+			"manual-entries", new String[0]);
+
+		if (ArrayUtil.isEmpty(manualEntries)) {
+			manualEntries = portletPreferences.getValues(
+				"manualEntries", new String[0]);
+		}
+
+		if (ArrayUtil.isEmpty(assetEntryXmls) &&
+			ArrayUtil.isNotEmpty(manualEntries)) {
+
+			assetEntryXmls = getAssetEntryXmls(manualEntries);
+
+			portletPreferences.setValues("asset-entry-xml", assetEntryXmls);
+		}
+
+		if (ArrayUtil.isNotEmpty(assetEntryXmls)) {
+			upgradeUuids(assetEntryXmls);
+
+			portletPreferences.setValues("assetEntryXml", assetEntryXmls);
+		}
+
 		return PortletPreferencesFactoryUtil.toXML(portletPreferences);
+	}
+
+	protected void upgradeUuids(String[] assetEntryXmls) throws Exception {
+		for (int i = 0; i < assetEntryXmls.length; i++) {
+			String assetEntry = assetEntryXmls[i];
+
+			Document document = SAXReaderUtil.read(assetEntry);
+
+			Element rootElement = document.getRootElement();
+
+			Element assetTypeElementUuid = rootElement.element(
+				"asset-entry-uuid");
+
+			String assetUuid = assetTypeElementUuid.getStringValue();
+
+			List<JournalArticle> articles = JournalArticleUtil.findByUuid(
+				assetUuid);
+
+			if (articles.size() > 0) {
+				JournalArticleResource resource =
+						JournalArticleResourceUtil.findByPrimaryKey(
+							articles.get(0).getResourcePrimKey());
+
+				rootElement.remove(assetTypeElementUuid);
+
+				assetTypeElementUuid.setText(resource.getUuid());
+
+				rootElement.add(assetTypeElementUuid);
+
+				document.setRootElement(rootElement);
+
+				assetEntryXmls[i] = document.formattedString(StringPool.BLANK);
+			}
+		}
 	}
 
 }
