+++ portal-impl/src/com/liferay/portal/upgrade/v6_0_3/UpgradeAssetPublisher.java
@@ -14,14 +14,19 @@
 
 package com.liferay.portal.upgrade.v6_0_3;
 
+import com.liferay.portal.kernel.dao.jdbc.DataAccess;
 import com.liferay.portal.kernel.upgrade.BaseUpgradePortletPreferences;
+import com.liferay.portal.kernel.util.GetterUtil;
 import com.liferay.portal.kernel.util.StringPool;
 import com.liferay.portal.kernel.util.Validator;
 import com.liferay.portal.kernel.xml.Document;
 import com.liferay.portal.kernel.xml.Element;
 import com.liferay.portal.kernel.xml.SAXReaderUtil;
-import com.liferay.portal.upgrade.util.UpgradeAssetPublisherManualEntries;
 import com.liferay.portlet.PortletPreferencesFactoryUtil;
+
+import java.sql.Connection;
+import java.sql.PreparedStatement;
+import java.sql.ResultSet;
 
 import javax.portlet.PortletPreferences;
 
@@ -29,6 +34,43 @@ import javax.portlet.PortletPreferences;
  * @author Julio Camarero
  */
 public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
+
+	public static void upgradeToAssetEntryUuidElement(Element rootElement)
+		throws Exception {
+
+		Element assetEntryIdElement = rootElement.element("assetEntryId");
+
+		long assetEntryId = GetterUtil.getLong(assetEntryIdElement.getText());
+
+		Connection con = null;
+		PreparedStatement ps = null;
+		ResultSet rs = null;
+
+		try {
+			con = DataAccess.getUpgradeOptimizedConnection();
+
+			ps = con.prepareStatement(
+				"select classUuid from AssetEntry where entryId = ?");
+
+			ps.setLong(1, assetEntryId);
+
+			rs = ps.executeQuery();
+
+			if (rs.next()) {
+				String classUuid = rs.getString("classUuid");
+
+				Element assetEntryUuidElement = rootElement.addElement(
+					"assetEntryUuid");
+
+				assetEntryUuidElement.addText(classUuid);
+
+				rootElement.remove(assetEntryIdElement);
+			}
+		}
+		finally {
+			DataAccess.cleanUp(con, ps, rs);
+		}
+	}
 
 	protected String[] getAssetEntryXmls(String[] assetEntryXmls)
 		throws Exception {
@@ -42,8 +84,7 @@ public class UpgradeAssetPublisher extends BaseUpgradePortletPreferences {
 
 			Element rootElement = document.getRootElement();
 
-			UpgradeAssetPublisherManualEntries.upgradeToAssetEntryUuidElement(
-				rootElement);
+			upgradeToAssetEntryUuidElement(rootElement);
 
 			newAssetEntryXmls[i] = document.formattedString(StringPool.BLANK);
 		}
